requirements.txt
djangorestframework==3.13.1
python-dateutil==2.8.2
parameterized==0.8.1
